ahora los proyectos tienen una lista propia de trabajos y objetivos

+++ tsServidor/gql/Objetivos.ts
@@ -22,7 +22,7 @@ export const typeDefs = gql`
        coords: Coords,
        angulo:Float,
        stuck:Boolean,
-       puntaje:Int,
+       puntaje:Float,
        centroMasa:Coords,
        nivel: Int,
        turnoNivel:Float,
